refactor(address-book): extract import-dialog + import-export lib
Moves the inline import-controls JSX into `_components/import-dialog.tsx`
(68 lines) and all file-parse/fetch/export logic into
`_lib/address-book-import-export.ts` (106 lines), shrinking
AddressBookClient from 393 → 277 lines. All 38 characterization tests pass.
